Refactored packages and reorganized setup file
